fix(shareScreen): in-correctly assigning localStreamSignal at the end
on the sender side we don't need localStream at all. we only want to send its tracks to the remote peer . previously we were overriding local stream at the end of share screen function it caused some issues. for example if we wanted to do a video call right after a share screen call , since we overrode the local stream our video track would be the screen share track ! to solve this we had to ask for user media again in video call function but now with this changes we no longer need to do so.